Refactor: Flatten STT compilation directory structure and sync READMEs
