+++ compilation/stt/requirements.txt
@@ -1,0 +1,5 @@
+transformers==4.50.0
+datasets==3.6.0
+librosa==0.11.0
+soundfile==0.13.1
+tqdm==4.67.1
